The new porch board speaks: 96 kbps on the card, a stop that really stops, one volume ceiling, and a speaker test on the desk and the phone remote
Board bring-up (v5.34–5.40, all flashed and confirmed on 10.27.27.247):
- castle_eink.h: three %u warnings fixed with casts; the pending QR and
  remote `.catch(sync)` patches applied; devices.toml names the new board.
- Two bitrates: the flash build keeps 32 kbps (its 2.9 MB partition), the
  card streams 96 kbps copies from audio/card/ (render_audio card_dir(),
  sd_sync prefers them). The render's card/test dirs follow a redirected
  OUT — a fixed path once let a sandboxed test sweep the real render.
- Mono stays: both amps are pinned LEFT and the S2 puts a mono stream in
  the left slot at full level (HARDWARE_FINDINGS §8 tells the whole audio
  diagnosis — it was the 5 V rail, not the chip).
- scene_stop stops the scene SCRIPTS (v5.35): Vigil's 30 s loop had been
  walking back on under every "quiet board" test; /api/play runs
  run_scene("halt") first (v5.37) so a song or a tone keeps the speakers.
- One ceiling, scenes.yaml hardware.audio.max_volume: generated into every
  scene's level and rig.h, clamped by /api/volume and the emulator, held
  equal by tests. 0.8 for the bare drivers, 1.0 for the shrouded pair.
- Speaker test: five tones (make audio → audio/test/, sd_sync tones) in the
  desk's 🏰 panel beside the strip test, and a speaker strip on /remote.
- castle.yaml split on its audio seam into castle_audio.yaml (500-line
  cap). v5.40's internal task stacks did not audibly help a scene start —
  open in docs/ISSUE-scene-start-audio.md, with A/B/C/D and the next tests.

+++ firmware/castle_eink.h
@@ -219,8 +219,10 @@ inline void render() {
   text(4, 54, line);
 
   const uint32_t up_min = (uint32_t) (esp_timer_get_time() / 60000000ULL);
-  snprintf(line, sizeof(line), "up    %ud %02u:%02u", up_min / 1440,
-           (up_min / 60) % 24, up_min % 60);
+  // uint32_t is `unsigned long` on xtensa, so %u needs the cast — the same
+  // idiom the boots/crashes line below already uses.
+  snprintf(line, sizeof(line), "up    %ud %02u:%02u", (unsigned) (up_min / 1440),
+           (unsigned) ((up_min / 60) % 24), (unsigned) (up_min % 60));
   text(4, 67, line);
 
   snprintf(line, sizeof(line), "boots %u  crashes %u",
